pyproject.toml
[project]
name = "portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-fasthtml>=0.12.33",
]
